Rename "Switchable 5200" cartridges to "Super Cart"
Bryan's cartridge PCB has the cartridge's name printed on it: "512K
Super Cart" - it is as official a name as we can get.

The term "Switchable" was so far used only in the context of "Switchable
XEGS" cartridges, referring to their ability to be switched on/off (i.e.
enabled/disabled). The 5200 Super Cart don't have this ability.

Shift the word "5200" in the cartridge names to after the size, to make
the names in line with other cartridges.