Hard-code list of objects in Android.mk.in
Since the build system was migrated to Automake, the configure script
does no longer provide the OBJS variable - the list of objects is
generated during processing of src/Makefile.in -> src/Makefile, and is
internal to that Makefile. There is apparently no feasible way to access
that list of objects when generating src/android/jni/Android.mk, so I
decided to simply hardcode the listof objects. It is not that bad of an
idea, since the Android package is supposed to be configured with a
specified set of features anyway (as described in DOC/README.android).

A better solution would involve ceasing from using ndk-build and
building the android target's C library using the same Automake scripts
as the other targets, but that would require a more serious rewrite.