Fix segfault when exiting monitor
If atari800 was started with "-monitor", then entering "quit" in the
monitor would continue emulation instead of quitting, because the result
of PLATFOM_Exit was ignored. This would result in segmentation fault,
because various SDL subsystems would be uininitialized at this point
(as the call to PLATFORM_Exit would close them and then not reopen).

Solve by replacing the call to PLATFORM_Exit with Atari800_Exit (because
PLATFORM_Exit was not closing enough subsystems anyway) and testing the
function's result.