Implemented delayed NMI handling
Altirra Hardware Reference Manual:

A taken relative branch delays interrupt acknowledgment by one cycle: a
case in which the earliest opportunity to respond to an interrupt is
immediately after the branch instead is delayed to the next instruction.
This occurs for any Bcc instruction which does not cross a page
boundary. The effect does not occur if the branch instruction crosses
a page (4 cycles), or for any other control flow instruction such as
JMP, JSR, RTS, or RTI.

Fixes #79.